pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cartographer"
version = "0.1.0"
description = "local-first knowledge filesystem"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
]

[project.scripts]
cart = "cartographer.cli:main"
cartog = "cartographer.cli:main"
cartographer = "cartographer.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cartographer*"]
